made click conditions less strict
will click more consistently when unfocused / decimal offset

i think one of the reasons why the clicks dont happen sometimes is because its too precise. which results in skipping if there so happens to be an offset in the values, or if there are frame drops due to reduces app speed when unfocused.

like it clicks at 10, but not at 10.1, as the modulus remainder for it becomes 0.1 rather than 0 (edited)

fixed by making isDivisible() use int for the value. and by converting the flipper values to int before passing it to the function.

(isdivisible() calls are used in the if condition to play clicks.)